new possible structure for flask_bridge in frontend
