+++ How_to_run.txt
@@ -11,7 +11,7 @@ Run node server, need cd in app
 1.
 cd clinic_hub
 2.
-node src/server.js
+node server/index.js
 
 to run all pages and servers, starts the node(rbac), flask request sender, need cd in app
 1.
