+++ clinic-hub/server/constants/roles.js
@@ -1,0 +1,7 @@
+const ROLES = {
+  ADMIN: "admin",
+  DOCTOR: "doctor",
+  RECEPTIONIST: "receptionist",
+};
+
+export default ROLES;
