+++ clinic-hub/server/middleware/authorization_middleware.js
@@ -1,9 +1,9 @@
-import jwt from 'jsonwebtoken';
+import jwt from "jsonwebtoken";
 // middleware/authorization_middleware.js
 const authorize = (allowedRoles = []) => {
   return (req, res, next) => {
-    const token = req.headers.authorization?.split(' ')[1];
-    
+    const token = req.headers.authorization?.split(" ")[1];
+
     if (!token) return res.status(401).json({ message: "No token" });
 
     try {
