--- clinic-hub/src/constants/roles.js
@@ -1,7 +1,0 @@
-const ROLES = {
-  ADMIN: 'admin',
-  DOCTOR: 'doctor',
-  RECEPTIONIST: 'receptionist'
-};
-
-export default ROLES;
